+++ crates/plugin_api/src/lib.rs
@@ -70,6 +70,8 @@ pub const API_VERSION: &str = concat!(env!("CARGO_PKG_VERSION"), "\0");
 /// The shared trait defining the Ewwii plugin API
 pub trait EwwiiAPI: Send + Sync {
     // == General Stuff == //
+    fn metadata_id(&self) -> String;
+
     /// Log a message from the host
     fn log(&self, msg: &str);
     /// Log a warning from the host